content-security-policy/to-port-from-blink/connect-src-beacon-redirect-to-blocked-expected.txt

Verify that a CSP connect-src directive blocks redirects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.sendBeacon("resources/redir.php?url=http://127.0.0.1:8080/navigation/resources/save-beacon.php%3Fname%3Dcsp-redirect-blocked", "ping"); is true
PASS Beacon sent successfully
PASS Content-Type: text/plain;charset=UTF-8
PASS Origin: http://127.0.0.1:8000
PASS Referer: http://127.0.0.1:8000/security/contentSecurityPolicy/connect-src-beacon-redirect-to-blocked.html
PASS Request-Method: POST
PASS Length: 4
PASS Body: ping
PASS 
PASS successfullyParsed is true

TEST COMPLETE
